+++ src/Adapters/Interfaces/NameInterface.php
@@ -25,7 +25,8 @@ interface NameInterface
      * Build a dynamic database name.
      *
      * @param string $dbNameHash The current db-name-hash based on the database-building file content,
-     *                           pre-migration-imports, migrations, seeder-settings, connection and transactions.
+     *                           database-name-prefix, pre-migration-imports, migrations, seeder-settings, connection
+     *                           and transactions.
      * @return string
      * @throws AdaptLaravelMySQLAdapterException Thrown when the database name is invalid.
      */
@@ -35,7 +36,7 @@ interface NameInterface
      * Generate the path (including filename) for the snapshot file.
      *
      * @param string $snapshotHash The current snapshot-hash based on the database-building file content,
-     *                             pre-migration-imports, migrations and seeder-settings.
+     *                             database-name-prefix, pre-migration-imports, migrations and seeder-settings.
      * @return string
      */
     public function generateSnapshotPath(string $snapshotHash): string;