+++ src/Adapters/LaravelMySQL/LaravelMySQLName.php
@@ -20,7 +20,8 @@ class LaravelMySQLName implements NameInterface
      * Build a dynamic database name.
      *
      * @param string $dbNameHash The current db-name-hash based on the database-building file content,
-     *                           pre-migration-imports, migrations, seeder-settings, connection and transactions.
+     *                           database-name-prefix, pre-migration-imports, migrations, seeder-settings, connection
+     *                           and transactions.
      * @return string
      * @throws AdaptLaravelMySQLAdapterException Thrown when the database name is invalid.
      */
@@ -36,7 +37,7 @@ class LaravelMySQLName implements NameInterface
      * Generate the path (including filename) for the snapshot file.
      *
      * @param string $snapshotHash The current snapshot-hash based on the database-building file content,
-     *                             pre-migration-imports, migrations and seeder-settings.
+     *                             database-name-prefix, pre-migration-imports, migrations and seeder-settings.
      * @return string
      */
     public function generateSnapshotPath(string $snapshotHash): string