+++ src/Support/Hasher.php
@@ -51,7 +51,7 @@ class Hasher
     }
 
     /**
-     * Build a hash based on the source files.
+     * Build a hash based on the source files (and the database name prefix).
      *
      * @return string
      * @throws AdaptConfigException Thrown when a directory or file could not be opened.
@@ -72,7 +72,7 @@ class Hasher
             $hashes[$path] = $this->di->filesystem->md5File($path);
         }
 
-        static::$sourceFilesHash = md5(serialize($hashes));
+        static::$sourceFilesHash = md5(serialize($hashes) . $this->config->databasePrefix);
 
         $this->di->log->info('Generated a hash of the database-related files', $logTimer);
 
@@ -193,7 +193,8 @@ class Hasher
     /**
      * Generate the scenario-hash based on the way this DatabaseBuilder will build this database.
      *
-     * Based on the database-building file content, pre-migration-imports, migrations and seeder-settings.
+     * Based on the database-building file content, database-name-prefix, pre-migration-imports, migrations and
+     * seeder-settings.
      *
      * @param string[] $seeders The seeders that will be run.
      * @return string
@@ -211,11 +212,11 @@ class Hasher
     /**
      * Generate a hash to use in the database name.
      *
-     * Based on the database-building file content, pre-migration-imports, migrations, seeder-settings, connection and
-     * transactions.
+     * Based on the database-building file content, database-name-prefix, pre-migration-imports, migrations,
+     * seeder-settings, connection and transactions.
      *
      * @param string[] $seeders          The seeders that will be run.
-     * @param string   $databaseModifier The modifier to use.
+     * @param string   $databaseModifier The modifier to use (eg. ParaTest suffix).
      * @return string
      */
     public function generateDBNameHash(array $seeders, string $databaseModifier): string