+++ tests/Integration/Laravel/DatabaseBuilder/ReuseDBTest.php
@@ -65,7 +65,7 @@ class ReuseDBTest extends LaravelTestCase
                     ->transactions(false)
                     ->isBrowserTest(false),
                 'updateReuseTableQuery' => "UPDATE `" . Settings::REUSE_TABLE . "` SET `inside_transaction` = 0",
-                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.3dd190-53855ffa7573.sqlite',
+                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.afd818-53855ffa7573.sqlite',
                 'expectedUserCount' => 0,
                 'expectedException' => null,
             ],
@@ -77,7 +77,7 @@ class ReuseDBTest extends LaravelTestCase
                     ->transactions(false)
                     ->isBrowserTest(false),
                 'updateReuseTableQuery' => "UPDATE `" . Settings::REUSE_TABLE . "` SET `inside_transaction` = 0",
-                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.3dd190-53855ffa7573.sqlite',
+                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.afd818-53855ffa7573.sqlite',
                 'expectedUserCount' => 0,
                 'expectedException' => null,
             ],
@@ -113,7 +113,7 @@ class ReuseDBTest extends LaravelTestCase
                     ->transactions(true)
                     ->isBrowserTest(false),
                 'updateReuseTableQuery' => "UPDATE `" . Settings::REUSE_TABLE . "` SET `inside_transaction` = 0",
-                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.3dd190-e86e4a953f0d.sqlite',
+                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.afd818-e86e4a953f0d.sqlite',
                 'expectedUserCount' => 0,
                 'expectedException' => null,
             ],
@@ -125,7 +125,7 @@ class ReuseDBTest extends LaravelTestCase
                     ->transactions(true)
                     ->isBrowserTest(false),
                 'updateReuseTableQuery' => "UPDATE `" . Settings::REUSE_TABLE . "` SET `inside_transaction` = 0",
-                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.3dd190-e86e4a953f0d.sqlite',
+                'expectedDBName' => $this->wsAdaptStorageDir . '/test-database.afd818-e86e4a953f0d.sqlite',
                 'expectedUserCount' => 1,
                 'expectedException' => null,
             ],