+++ tests/Integration/Laravel/DatabaseBuilder/SnapshotTest.php
@@ -46,7 +46,7 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, false)
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -57,7 +57,7 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, false, true)
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -68,8 +68,8 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -81,7 +81,7 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, false)
                     ->seeders([]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -101,7 +101,7 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -114,7 +114,7 @@ class SnapshotTest extends LaravelTestCase
                     ->preMigrationImports(['sqlite' => $this->wsPreMigrationsDir . '/pre-migration-import-1.sqlite'])
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-abda22425307.sqlite',
+                    'snapshot.database.afd818-abda22425307.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -126,7 +126,7 @@ class SnapshotTest extends LaravelTestCase
                     ->preMigrationImports(['sqlite' => $this->wsPreMigrationsDir . '/pre-migration-import-1.sqlite'])
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-0a7748252665.sqlite',
+                    'snapshot.database.afd818-0a7748252665.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -138,8 +138,8 @@ class SnapshotTest extends LaravelTestCase
                     ->preMigrationImports(['sqlite' => $this->wsPreMigrationsDir . '/pre-migration-import-1.sqlite'])
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-0a7748252665.sqlite',
-                    'snapshot.database.3dd190-abda22425307.sqlite',
+                    'snapshot.database.afd818-0a7748252665.sqlite',
+                    'snapshot.database.afd818-abda22425307.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -152,7 +152,7 @@ class SnapshotTest extends LaravelTestCase
                     ->preMigrationImports(['sqlite' => $this->wsPreMigrationsDir . '/pre-migration-import-1.sqlite'])
                     ->seeders([]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-abda22425307.sqlite',
+                    'snapshot.database.afd818-abda22425307.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -174,7 +174,7 @@ class SnapshotTest extends LaravelTestCase
                     ->preMigrationImports(['sqlite' => $this->wsPreMigrationsDir . '/pre-migration-import-1.sqlite'])
                     ->seeders([]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-abda22425307.sqlite',
+                    'snapshot.database.afd818-abda22425307.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => true,
@@ -186,8 +186,8 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => false,
@@ -198,9 +198,9 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([UserSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-5f8175d30493.sqlite',
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-5f8175d30493.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
                 'expectedDatabase' => null,
                 'removeAdaptStorageDir' => false,
@@ -213,10 +213,10 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
-                'expectedDatabase' => 'test-database.3dd190-3e4b86d50da4-1.sqlite',
+                'expectedDatabase' => 'test-database.afd818-3e4b86d50da4-1.sqlite',
                 'removeAdaptStorageDir' => true,
                 'expectUsers' => [],
             ],
@@ -227,10 +227,10 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([DatabaseSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
-                'expectedDatabase' => 'test-database.3dd190-3e4b86d50da4-1.sqlite',
+                'expectedDatabase' => 'test-database.afd818-3e4b86d50da4-1.sqlite',
                 'removeAdaptStorageDir' => false,
                 'expectUsers' => ['imported-snapshot-after-seeders'],
             ],
@@ -240,11 +240,11 @@ class SnapshotTest extends LaravelTestCase
                     ->snapshots(true, true, true)
                     ->seeders([UserSeeder::class]),
                 'expectedSnapshots' => [
-                    'snapshot.database.3dd190-5f8175d30493.sqlite',
-                    'snapshot.database.3dd190-8bd51f9f0b21.sqlite',
-                    'snapshot.database.3dd190-a34cd538e35f.sqlite',
+                    'snapshot.database.afd818-5f8175d30493.sqlite',
+                    'snapshot.database.afd818-8bd51f9f0b21.sqlite',
+                    'snapshot.database.afd818-a34cd538e35f.sqlite',
                 ],
-                'expectedDatabase' => 'test-database.3dd190-7a4fbda6d161-1.sqlite',
+                'expectedDatabase' => 'test-database.afd818-7a4fbda6d161-1.sqlite',
                 'removeAdaptStorageDir' => false,
                 'expectUsers' => ['imported-snapshot-before-seeders'],
             ],
@@ -303,4 +303,48 @@ class SnapshotTest extends LaravelTestCase
             $this->assertSame('1', $row[0]->total);
         }
     }
+
+    /**
+     * Build the two snapshot files that go into tests/workspaces/scenario1/database/adapt-test-storage
+     *
+     * Steps:
+     * - delete the snapshot.database.x.sqlite files from tests/workspaces/scenario1/database/adapt-test-storage/
+     * - run this test
+     * - copy the tests/workspaces/scenario1/database/adapt-test-storage/snapshot.database.x.sqlite files to dir above.
+     */
+/*
+    public function test_build_snapshot_sqlite_databases()
+    {
+        $this->prepareWorkspace("$this->workspaceBaseDir/scenario1", $this->wsCurrentDir, true);
+
+        $config = $this->newConfigDTO('sqlite')
+            ->snapshots(true, true, true)
+            ->seeders([DatabaseSeeder::class]);
+
+        // build the database
+        $this->newDatabaseBuilder($config)->execute();
+
+
+
+        // find the snapshot files
+        $snapshotFiles = collect((array) scandir($this->wsAdaptStorageDir))->filter(function ($path) use ($config) {
+            return preg_match('/^'.preg_quote($config->snapshotPrefix).'/', $path);
+        });
+
+        foreach ($snapshotFiles as $snapshotFile) {
+
+            DB::connection($config->connection)->disconnect();
+            config([
+                'database.connections.' . $config->connection . '.database' => "$this->wsAdaptStorageDir/$snapshotFile"
+            ]);
+
+            $userCount = DB::connection($config->connection)->select("SELECT COUNT(*) AS total FROM `users`")[0]->total;
+
+            DB::connection($config->connection)->insert(
+                "INSERT INTO `users` (`username`) VALUES (:username)",
+                ['username' => $userCount ? 'imported-snapshot-after-seeders' : 'imported-snapshot-before-seeders']
+            );
+        }
+    }
+*/
 }