+++ tests/workspaces/scenario1/config/code_distortion.adapt.php
@@ -9,11 +9,11 @@ return [
      |
      | You might share your database server between different projects. If so
      | then you can specify a unique project name here to ensure that Adapt
-     | doesn't interfere with Adapt test-databases in other projects.
+     | doesn't interfere with Adapt test-databases from other projects.
      |
      */
 
-    'project-name' => env('ADAPT_PROJECT_NAME', ''),
+    'project_name' => env('ADAPT_PROJECT_NAME', ''),
 
     /*
     |--------------------------------------------------------------------------
@@ -72,7 +72,7 @@ return [
     |
     */
 
-    'pre-migration-imports' => [
+    'pre_migration_imports' => [
         'mysql' => [],
         'sqlite' => [],
         'pgsql' => [],
@@ -83,9 +83,9 @@ return [
     | Reuse Test-Databases
     |--------------------------------------------------------------------------
     |
-    | When a test-database already exists and was left in a clean state,
-    | it can be reused without needing to rebuild it, saving time.
-    | This is best used with the create-test-dbs setting below.
+    | When a test-database already exists (and was left in a clean state),
+    | it can be reused without Adapt needing to rebuild it, saving time.
+    | This is best used with the dynamic_test_dbs setting below.
     |
     | This config setting can be overridden by adding the
     | $reuseTestDBs property to your test-class.
@@ -94,7 +94,7 @@ return [
     |
     */
 
-    'reuse-test-dbs' => env('ADAPT_REUSE_TEST_DBS', true),
+    'reuse_test_dbs' => env('ADAPT_REUSE_TEST_DBS', true),
 
     /*
     |--------------------------------------------------------------------------
@@ -103,37 +103,36 @@ return [
     |
     | A new database (based on your original database name) will be created
     | for each scenario your tests need. This is best used with the
-    | reuse-test-dbs setting above. These dbs are safe to delete.
+    | reuse_test_dbs setting above. These dbs are safe to delete.
     |
-    | eg. "your_database_name_d42a95e7d1afcb4f9902677f74e9b2d5"
+    | An dynamic database will be called something like:
+    | "test_your_database_name_17bd3c_d266ab43ac75"
     |
-    | This config setting can be overridden by adding the
-    | $dynamicTestDBs property to your test-class.
+    | This config setting can be overridden by adding the $dynamicTestDBs
+    | property to your test-class.
     |
-    | NOTE: You should turn this off when browser testing (eg. Dusk). This
-    | setting can be overridden by adding the $dynamicTestDBs property
-    | to your test-class.
+    | NOTE: This is turned off automatically when browser testing (eg. Dusk).
     |
     | eg.
     | protected bool $dynamicTestDBs = true;
     |
     */
 
-    'dynamic-test-dbs' => env('ADAPT_DYNAMIC_TEST_DBS', true),
+    'dynamic_test_dbs' => env('ADAPT_DYNAMIC_TEST_DBS', true),
 
     /*
     |--------------------------------------------------------------------------
     | Use Transactions
     |--------------------------------------------------------------------------
     |
-    | Your tests can run within a transaction that's rolled-back afterwards,
-    | leaving the database fresh so it won't need rebuilding for the next
-    | test.
+    | Your tests can run within a transaction that's rolled-back afterwards.
+    | This leaves the database fresh so it won't need to be rebuilt
+    | during the next test-run.
     |
-    | NOTE: You should turn this off when browser testing (eg. Dusk).
+    | This config setting can be overridden by adding the $transactions
+    | property to your test-class.
     |
-    | This config setting can be overridden by adding the
-    | $transactions property to your test-class.
+    | NOTE: This is turned off automatically when browser testing (eg. Dusk).
     |
     | eg.
     | protected bool $transactions = true;
@@ -149,10 +148,13 @@ return [
     |
     | Database dumps/copies can be taken of each test-db, which are imported
     | automatically when needed saving migration + seeding time. Snapshot
-    | files are stored in the "storage-dir" and can be removed safely.
+    | files are stored in the "storage_dir" and can be removed safely.
     |
     | If you use Adapt to seed many tests differently, it may be worth
     | also taking a snapshot 'after-migrations' (ie. before seeding).
+    |
+    | Taking snapshots after seeding is useful when performing
+    | browser tests (eg. Dusk).
     |
     | These config settings can be overridden by adding the $snapshotsEnabled,
     | $takeSnapshotAfterMigrations and $takeSnapshotAfterSeeders properties
@@ -161,9 +163,9 @@ return [
     */
 
     'snapshots' => [
-        'enabled' => env('ADAPT_USE_SNAPSHOTS', true),
-        'take-after-migrations' => env('ADAPT_TAKE_SNAPSHOTS_AFTER_MIGRATIONS', false),
-        'take-after-seeders' => env('ADAPT_TAKE_SNAPSHOTS_AFTER_SEEDERS', true),
+        'enabled' => env('ADAPT_USE_SNAPSHOTS', false),
+        'take_after_migrations' => env('ADAPT_TAKE_SNAPSHOTS_AFTER_MIGRATIONS', false),
+        'take_after_seeders' => env('ADAPT_TAKE_SNAPSHOTS_AFTER_SEEDERS', true),
     ],
 
     /*
@@ -177,9 +179,7 @@ return [
     |
     */
 
-    'storage-dir' => realpath(
-        base_path('../../../../tests/workspaces/current/database')
-    ) . '/adapt-test-storage',
+    'storage_dir' => realpath(base_path('../../../../tests/workspaces/current/database')) . '/adapt-test-storage',
 
     /*
     |--------------------------------------------------------------------------
@@ -192,10 +192,11 @@ return [
     |
     */
 
-    'look-for-changes-in' => [
-        database_path('factories'),
-        database_path('migrations'),
-        database_path('seeds'),
+    'look_for_changes_in' => [
+        realpath(base_path('../../../../tests/workspaces/current/database')) . '/factories',
+        realpath(base_path('../../../../tests/workspaces/current/database')) . '/migrations',
+//        realpath(base_path('../../../../tests/workspaces/current/database')) . '/seeders',     // Laravel 8 and after
+        realpath(base_path('../../../../tests/workspaces/current/database')) . '/seeds',       // before Laravel 8
     ],
 
     /*
@@ -220,7 +221,7 @@ return [
     | '!mysql < sqlite'
     */
 
-    'remap-connections' => env('ADAPT_REMAP_CONNECTIONS', ''),
+    'remap_connections' => env('ADAPT_REMAP_CONNECTIONS', ''),
 
     /*
      |--------------------------------------------------------------------------